fix(Config): Remove work method in favor of configure
All configuration should be done in the `configure` method.
Code previously found in the `work` method can safely be
moved to the bottom of the `configure` method.  Workers will
still only be activated if the `registerWorkers` setting
is set to `true`. This change simplifies environment
overrides.

BREAKING CHANGE: Remove `work` method in favor of `configure`.