Update AbstractQueueProvider.cfc
I think this will fix the issue of the queue not being passed along in the chain when it dispatches the next job. I noticed that the connection was also missing, so I added it. I tested with the queue, and it works for me. I did not try testing the connection. This works if you create a chain and specify a queue, and all the specified queues are the same. There is still an issue where if you mix and match queue names, it will not use the correct queue the first time it comes to the new queue name. It won't create the new job with the correct queue at that point, but rather use the queue of the prior job in the chain.