fix(DBProvider): Fix duplicate job runs
Duplicate job runs could occur because we brought back many job records
and added them to the executor while only marking the currently running
jobs as reserved.  Instead of just marking them as reserved right away
(which we do now), we further improved the system by only fetching
an amount of job records equal to the current capacity of the Worker Pool.
Finally, when a job finishes we immediately fetch more job records as
to not waste time waiting for the next scheduled fetch.