feat(AbstractQueueProvider): Pass in the full job object to `push`
This allows the provider to determine the best way to serialize the job,
instead of just using `serializeJSON`. (The provider will need to pass
the memento of the job to `deserializeJob` if using that helper still.)

BREAKING CHANGE: The message signature of `push` has changed. Third-party
providers need to update to the new message signature.